# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations/{id}/users": {
      "get": {
        "operationId": "OrganizationsController_getDepartmentUsers",
        "summary": "获取部门用户列表",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "部门ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取部门用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "avatar": {
                            "type": "string"
                          },
                          "dutyName": {
                            "type": "string"
                          },
                          "employeeId": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/documents": {
      "post": {
        "operationId": "DocumentsController_createDocument",
        "summary": "创建文档",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DocumentsController_getDocuments",
        "summary": "获取文档列表",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_getDocument",
        "summary": "获取文档详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/documents/{id}/collaborators": {
      "post": {
        "operationId": "DocumentsController_addCollaborator",
        "summary": "添加协作者",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCollaboratorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Chat API",
    "description": "The chat API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "文档标题"
          },
          "type": {
            "type": "string",
            "description": "文档类型",
            "enum": [
              "word",
              "excel"
            ]
          }
        },
        "required": [
          "title",
          "type"
        ]
      },
      "AddCollaboratorDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "用户ID"
          },
          "role": {
            "type": "string",
            "description": "协作者角色",
            "enum": [
              "editor",
              "viewer"
            ]
          }
        },
        "required": [
          "userId",
          "role"
        ]
      }
    }
  }
}